Features/Bao

+++ FootballFieldManagement/FootballFieldManagement/ViewModels/GoodsViewModel.cs
@@ -41,8 +41,6 @@ namespace FootballFieldManagement.ViewModels
         public ICommand ExitImportCommand { get; set; } //thoát khỏi ImportGoodsWindow
         public ICommand CalculateTotalCommand { get; set; } //tính tổng tiền
 
-        //PayWindow
-        public ICommand PickGoodsCommand { get; set; } // Chọn 1 hàng 
         public GoodsViewModel()
         {
             //GoodsControl
@@ -60,8 +58,6 @@ namespace FootballFieldManagement.ViewModels
             ExitImportCommand = new RelayCommand<ImportGoodsWindow>((parameter) => true, (parameter) => parameter.Close());
             CalculateTotalCommand = new RelayCommand<ImportGoodsWindow>((parameter) => true, (parameter) => CalculateTotal(parameter));
 
-            //PayWindow
-            PickGoodsCommand = new RelayCommand<SellGoodsControl>((parameter) => true, (parameter) => BuyGoods(parameter));
         }
 
         //GoodsControl
@@ -148,23 +144,28 @@ namespace FootballFieldManagement.ViewModels
         }
         public void DeleteGoods(TextBlock txb)
         {
-            string idGoods = txb.Text;
-            List<string> idStockReceiptList = StockReceiptInfoDAL.Instance.QueryIdStockReceipt(idGoods);
+            MessageBoxResult result = MessageBox.Show("Xác nhận xóa hàng hóa?", "Thông báo", MessageBoxButton.YesNo, MessageBoxImage.Question);
 
-            bool isSuccessed1 = StockReceiptInfoDAL.Instance.DeleteFromDB(idGoods);
-            bool isSuccessed2 = true;
-            foreach (var idStockReceipt in idStockReceiptList)
+            if (result == MessageBoxResult.Yes)
             {
-                isSuccessed2 = StockReceiptDAL.Instance.DeleteFromDB(idStockReceipt);
-            }
-            bool isSuccessed3 = GoodsDAL.Instance.DeleteFromDB(idGoods);
-            if (isSuccessed1 && isSuccessed2 && isSuccessed3 || isSuccessed3)
-            {
-                MessageBox.Show("Xoá thành công!");
-            }
-            else
-            {
-                MessageBox.Show("Thực hiện thất bại!");
+                string idGoods = txb.Text;
+                List<string> idStockReceiptList = StockReceiptInfoDAL.Instance.QueryIdStockReceipt(idGoods);
+
+                bool isSuccessed1 = StockReceiptInfoDAL.Instance.DeleteFromDB(idGoods);
+                bool isSuccessed2 = true;
+                foreach (var idStockReceipt in idStockReceiptList)
+                {
+                    isSuccessed2 = StockReceiptDAL.Instance.DeleteFromDB(idStockReceipt);
+                }
+                bool isSuccessed3 = GoodsDAL.Instance.DeleteFromDB(idGoods);
+                if (isSuccessed1 && isSuccessed2 && isSuccessed3 || isSuccessed3)
+                {
+                    MessageBox.Show("Xoá thành công!");
+                }
+                else
+                {
+                    MessageBox.Show("Thực hiện thất bại!");
+                }
             }
         }
 
@@ -306,40 +307,5 @@ namespace FootballFieldManagement.ViewModels
             parameter.txtTotal.Text = (importPriceTmp * quantityTmp).ToString();
         }
 
-        //PayWindow
-        public void BuyGoods(SellGoodsControl parameter)
-        {
-            bool isExist = false;
-            if (GoodsDAL.Instance.GetGood(parameter.txbId.Text).Quantity == 0)
-            {
-                MessageBox.Show("Đã hết hàng!");
-                return;
-            }
-            List<BillInfo> billInfos = BillInfoDAL.Instance.ConvertDBToList();
-            foreach (var billInfo in billInfos)
-            {
-                if (billInfo.IdBill == int.Parse(parameter.txbIdBill.Text) && billInfo.IdGoods == int.Parse(parameter.txbId.Text))
-                {
-                    isExist = true;
-                    billInfo.Quantity += 1;
-                    if (billInfo.Quantity > GoodsDAL.Instance.GetGood(billInfo.IdGoods.ToString()).Quantity)
-                    {
-                        billInfo.Quantity -= 1;
-                        MessageBox.Show("Đạt số lượng hàng tối đa!");
-                        return;
-                    }
-                    if (BillInfoDAL.Instance.UpdateOnDB(billInfo))
-                    {
-                        MessageBox.Show("Đã thêm!");
-                    }
-                    return;
-                }
-            }
-            if (!isExist)
-            {
-                BillInfo billInfo = new BillInfo(int.Parse(parameter.txbIdBill.Text), int.Parse(parameter.txbId.Text), 1);
-                BillInfoDAL.Instance.AddIntoDB(billInfo);
-            }
-        }
     }
 }
